refactor(starknet_l1_provider): move out of lib.rs
Move things out of lib.rs, no logic changes
