+++ crates/starknet_sequencer_node/src/components.rs
@@ -6,9 +6,8 @@ use starknet_class_manager_types::EmptyClassManagerClient;
 use starknet_consensus_manager::consensus_manager::ConsensusManager;
 use starknet_gateway::gateway::{create_gateway, Gateway};
 use starknet_http_server::http_server::{create_http_server, HttpServer};
-use starknet_l1_provider::event_identifiers_to_track;
 use starknet_l1_provider::l1_provider::{create_l1_provider, L1Provider};
-use starknet_l1_provider::l1_scraper::L1Scraper;
+use starknet_l1_provider::l1_scraper::{event_identifiers_to_track, L1Scraper};
 use starknet_mempool::communication::{create_mempool, MempoolCommunicationWrapper};
 use starknet_mempool_p2p::create_p2p_propagator_and_runner;
 use starknet_mempool_p2p::propagator::MempoolP2pPropagator;
