+++ crates/starknet_sequencer_node/src/config/node_config.rs
@@ -24,8 +24,8 @@ use starknet_consensus_manager::config::ConsensusManagerConfig;
 use starknet_gateway::config::GatewayConfig;
 use starknet_http_server::config::HttpServerConfig;
 use starknet_infra_utils::path::resolve_project_relative_path;
+use starknet_l1_provider::l1_provider::L1ProviderConfig;
 use starknet_l1_provider::l1_scraper::L1ScraperConfig;
-use starknet_l1_provider::L1ProviderConfig;
 use starknet_mempool_p2p::config::MempoolP2pConfig;
 use starknet_monitoring_endpoint::config::MonitoringEndpointConfig;
 use starknet_sierra_multicompile::config::SierraCompilationConfig;
